config.py: Add 'Zulip' Option for ROOMS_TO_JOIN
Added this code which if the `BACKEND` is run as `'Zulip'`
`ROOMS_TO_JOIN` is updated accordingly.
I noticed there's not much code for Zulip(a cursory
search shows no matches in the repo). So this
will probably make someone's life easier down
the line when they port the bot.

closes #635